+++ CHANGES.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [0.6.0]
 
 - Adds `cookie` middleware.
+- New `Request::isMethod`.
 
 ## [0.5.0] - 2019-02-21